requirements.txt
# Core
Django>=4.2,<5.0

# Django REST Framework (for API)
djangorestframework>=3.14

# Environment variables
python-decouple>=3.8

# Database (PostgreSQL for Render)
psycopg2-binary>=2.9

# Production server
gunicorn>=21.2

# Static files handling
whitenoise>=6.6

# Security & utilities
django-cors-headers>=4.3

# Optional but recommended
requests>=2.31
